Canvas: ScanLine rendering, drawPixel, tilemap engine; review fixes
LoxCanvas:
- Replace Canvas.Pixels with ScanLine direct memory access ({$POINTERMATH ON})
- Add precomputed FCurrentPixel Cardinal for single-word pixel writes
- Add drawPixel(x,y) native using ScanLine for particle/starfield effects
- Add nearest-neighbor scaling (SetStretchBltMode COLORONCOLOR)
- Add tilemap system: createTilemap, setTile, drawTilemap with scroll
- Clear sprites and tilemaps in RegisterCanvasNatives between runs
- TTilemap class with visible-range culling for efficient rendering

LoxSound:
- Fix Int64 overflow in playTone square-wave toggle (i*freq*2)

Samples:
- Add scrolling starfield to space_invaders.lox using drawPixel
- Add tilemap_demo.lox: scrolling platformer with 7 tile types

+++ LoxSound.pas
@@ -143,7 +143,7 @@ begin
   SetLength(buf, n * 2);
   for i := 0 to n - 1 do
   begin
-    if ((i * freq * 2) div SAMPLE_RATE) mod 2 = 0 then
+    if ((Int64(i) * freq * 2) div SAMPLE_RATE) mod 2 = 0 then
       sv := 10000
     else
       sv := -10000;
